requirements.txt
arrow==1.2.3
asgiref==3.6.0
Babel==2.11.0
binaryornot==0.4.4
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
cookiecutter==2.1.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.5
django-autoslug==1.9.8
django-browser-reload==1.6.0
django-ckeditor==6.5.1
django-cleanup==6.0.0
django-import-export==3.0.2
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-phonenumber-field==7.0.2
django-tailwind==3.4.0
djangorestframework==3.14.0
docopt==0.6.2
et-xmlfile==1.1.0
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupPy==1.14
MarkupSafe==2.1.2
mod-wsgi==4.9.4
odfpy==1.4.1
openpyxl==3.0.10
phonenumbers==8.13.4
Pillow==9.4.0
pipreqs==0.4.11
python-dateutil==2.8.2
python-slugify==7.0.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tablib==3.3.0
tailwind==3.1.5b0
text-unidecode==1.3
urllib3==1.26.14
xlrd==2.0.1
xlwt==1.3.0
yarg==0.1.9
zope.event==4.6
zope.interface==5.5.2
